0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -16,7 +16,7 @@ size_t print_list(const list_t *h)
 		if (!h->str)
 			printf("[0] (nil)\n");
 		else
-			printf("[%lu] %s\n", (unsigned long)h->len, h->str);
+			printf("[%lu] %s\n", (unsigned int)h->len, h->str);
 		h = h->next;
 		val++;
 	}
